Hacked 1st and 2nd repeats
VexTab unfortunately doesn't support first and second repeats.  So, we resort to hacking a poor but nevertheless workable version by co-opting annotations set above the score.  These are associated with the (first) note rather than the bar itself.  The drawbacks are that a proper image of the repeat type is impossible with text and any bona fide annotation for the first note will take precedence.